Replace domain_forwarding
The behavior of domain_forwarding is somewhat unpredictable. It can be used to populate empty sets, but not to expand existing sets - instead in the cases it laxes domain restriction by expanding the internal domain mapping specific to the parameter object which forwards the domain. This in effect creates an inconsistency in how a set is handled between objects which in the extreme case can lead to some nasty and timeconsuming debugging. I am currently not sure how domain_forwarding interacts with objects with multiple declared domains, but it is probably safest to avoid when possible.